+++ lib/ui/pages/add_friend.dart
@@ -1,6 +1,6 @@
+import 'package:chat_flutter/ui/molecules/add_friend_list_tile.dart';
 import 'package:flutter/material.dart';
 import 'package:chat_flutter/model/user.dart';
-import 'add_friend_list_tile.dart';
 
 class AddFriendPage extends StatelessWidget {
   final TextEditingController _textEditingController = TextEditingController();
